structure refactor of tasks components
