+++ src/app/tasks/page.tsx
@@ -1,4 +1,5 @@
-import { CreateTask, TasksList } from '@/components'
+import { CreateTask } from './components/create-task'
+import { TasksList } from './components/tasks-list'
 
 export default function Page() {
 	return (
