+++ src/components/index.ts
@@ -5,5 +5,3 @@ export * from './loading/spinner'
 export * from './nav/nav'
 export * from './not-found/not-found'
 export * from './swr-provider/swr-provider'
-export * from './tasks/create-task'
-export * from './tasks/tasks-list'
